Make Lift simpler
Change Arm class
